+++ CHANGES.rst
@@ -93,6 +93,13 @@ residual_fringe
 - Use scipy.interpolate.BSpline instead of astropy.modeling.Spline1D in
   residual_fringe fitting utils [#7764]
 
+undersampling_correction
+------------------------
+
+- Changed default signal threshold, added efficient routine to flag neighborhood
+  pixels, added new unit test, improved earlier unit tests, updated docs. [#7740]
+  
+
 1.11.4 (2023-08-14)
 ===================